+++ cdrs-tokio/examples/generic_connection.rs
@@ -77,11 +77,9 @@ impl GenericClusterConfig<TransportTcp, TcpConnectionManager> for VirtualCluster
         // create a connection manager that points at the rewritten address so that's where it connects, but
         // then return a manager with the 'virtual' address for internal purposes.
         Ok(TcpConnectionManager::new(
-            NodeTcpConfigBuilder::new(
-                addr.rewrite(&self.mask, &self.actual),
-                self.authenticator.clone(),
-            )
-            .build(),
+            NodeTcpConfigBuilder::new(addr.rewrite(&self.mask, &self.actual))
+                .with_authenticator_provider(self.authenticator.clone())
+                .build(),
             self.keyspace_holder.clone(),
             Compression::None,
             None,