+++ cdrs-tokio/src/cluster/config_tcp.rs
@@ -1,41 +1,48 @@
 use std::net::SocketAddr;
 use std::sync::Arc;
 
-use crate::authenticators::SaslAuthenticatorProvider;
+use crate::authenticators::{NoneAuthenticatorProvider, SaslAuthenticatorProvider};
 
 /// Cluster configuration that holds per node TCP configs
+#[derive(Clone, Default)]
 pub struct ClusterTcpConfig(pub Vec<NodeTcpConfig>);
 
 /// Single node TCP connection config.
 #[derive(Clone)]
 pub struct NodeTcpConfig {
     pub addr: SocketAddr,
-    pub authenticator: Arc<dyn SaslAuthenticatorProvider + Send + Sync>,
+    pub authenticator_provider: Arc<dyn SaslAuthenticatorProvider + Send + Sync>,
 }
 
 /// Builder structure that helps to configure TCP connection for node.
 pub struct NodeTcpConfigBuilder {
     addr: SocketAddr,
-    authenticator: Arc<dyn SaslAuthenticatorProvider + Send + Sync>,
+    authenticator_provider: Arc<dyn SaslAuthenticatorProvider + Send + Sync>,
 }
 
 impl NodeTcpConfigBuilder {
-    pub fn new(
-        addr: SocketAddr,
-        authenticator: Arc<dyn SaslAuthenticatorProvider + Send + Sync>,
-    ) -> NodeTcpConfigBuilder {
+    pub fn new(addr: SocketAddr) -> NodeTcpConfigBuilder {
         NodeTcpConfigBuilder {
             addr,
-            authenticator,
+            authenticator_provider: Arc::new(NoneAuthenticatorProvider),
         }
     }
 
     /// Sets new authenticator.
+    #[deprecated(note = "Use with_authenticator_provider().")]
     pub fn authenticator(
-        mut self,
+        self,
         authenticator: Arc<dyn SaslAuthenticatorProvider + Send + Sync>,
     ) -> Self {
-        self.authenticator = authenticator;
+        self.with_authenticator_provider(authenticator)
+    }
+
+    /// Sets new authenticator.
+    pub fn with_authenticator_provider(
+        mut self,
+        authenticator_provider: Arc<dyn SaslAuthenticatorProvider + Send + Sync>,
+    ) -> Self {
+        self.authenticator_provider = authenticator_provider;
         self
     }
 
@@ -43,7 +50,7 @@ impl NodeTcpConfigBuilder {
     pub fn build(self) -> NodeTcpConfig {
         NodeTcpConfig {
             addr: self.addr,
-            authenticator: self.authenticator,
+            authenticator_provider: self.authenticator_provider,
         }
     }
 }