+++ cdrs-tokio/src/cluster/rustls_connection_manager.rs
@@ -99,7 +99,7 @@ impl RustlsConnectionManager {
 
         startup(
             &transport,
-            self.config.authenticator.deref(),
+            self.config.authenticator_provider.deref(),
             self.keyspace_holder.deref(),
             self.compression,
         )